when there is no poison, make sure the empty poison index is type int64

+++ armory/scenarios/poison.py
@@ -199,7 +199,7 @@ class Poison(Scenario):
             self.x_poison, self.y_poison, self.poison_index = (
                 self.x_clean,
                 self.y_clean,
-                np.array([]),
+                np.array([], dtype=np.int64),
             )
 
         self.record_poison_and_data_info()
